add module test support
